# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/members:
    get:
      operationId: listProjectMembers
      tags: [projects]
      summary: List project members
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectMember"
    post:
      operationId: addProjectMember
      tags: [projects]
      summary: Invite a user to the project by email (owner only)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteInput"
      responses:
        "200":
          description: Invite result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteResult"
  /projects/{projectId}/members/{userId}:
    delete:
      operationId: removeProjectMember
      tags: [projects]
      summary: Remove a member, or leave the project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
...
  /projects/{projectId}/messages:
    get:
      operationId: listProjectMessages
      tags: [projects]
      summary: Poll project chat messages (supports ?after=&limit= query params)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectChatMessage"
    post:
      operationId: sendProjectMessage
      tags: [projects]
      summary: Send a chat message (mention @ai for an AI reply)
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewChatMessageInput"
      responses:
        "201":
          description: Sent message (and optional AI reply)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatPostResponse"
...
components:
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    InviteInput:
      type: object
      properties:
        email:
          type: string
      required: [email]
...
    ProjectMember:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        addedAt:
          type: string
      required: [id, email, role, addedAt]
    InviteResult:
      type: object
      properties:
        found:
          type: boolean
        pending:
          type: boolean
        added:
          type: boolean
        id:
          type: integer
        email:
          type: string
        role:
          type: string
      required: [found]
...
    ProjectChatMessage:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        createdAt:
          type: string
        senderId:
          type: integer
          nullable: true
        isAi:
          type: boolean
        senderEmail:
          type: string
          nullable: true
      required: [id, content, createdAt, isAi]
    NewChatMessageInput:
      type: object
      properties:
        content:
          type: string
      required: [content]
    ChatPostResponse:
      type: object
      properties:
        message:
          $ref: "#/components/schemas/ProjectChatMessage"
        aiReply:
          $ref: "#/components/schemas/ProjectChatMessage"
          nullable: true
      required: [message]