# lib/api-zod/src/generated/types/chatPostResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectChatMessage } from './projectChatMessage';

export interface ChatPostResponse {
  message: ProjectChatMessage;
  aiReply?: ProjectChatMessage | null;
}