# lib/api-zod/src/generated/types/inviteResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InviteResult {
  found: boolean;
  pending?: boolean;
  added?: boolean;
  id?: number;
  email?: string;
  role?: string;
}