# lib/api-zod/src/generated/types/projectChatMessage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectChatMessage {
  id: number;
  content: string;
  createdAt: string;
  senderId?: number | null;
  isAi: boolean;
  senderEmail?: string | null;
}